Added basic primitive values
Primitive values are the inspiration for the entire project. The valueOf() function exists for every object in JavaScript and, since everything is an object, including numbers and strings (and even NaN), except only undefined and null as far as I know, I can make it so that before drawing I call valueOf(), in which case a number will return itself but a class that has overloaded valueOf() can return whatever while staying referenced. This means you can set multiple properties of any object to a primitive instance and change the returned value of all the properties at the same time with only one execution